Plant_disease/mlops/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
mlflow>=2.5.0
pandas>=1.5.0
psutil>=5.9.0
plotly>=5.13.0
pytest>=7.0.0
pytest-cov>=4.0.0
pre-commit>=3.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pillow>=9.4.0
tensorboard>=2.12.0
tqdm>=4.65.0QT